refactor: change project component structure
