+++ src/domain/member/components/index.js
@@ -1,4 +1,1 @@
 export { default as OnboardingItem } from "./OnboardingItem";
-export { default as KakaoLoginButton } from "./KakaoLoginButton";
-export { default as AppleLoginButton } from "./AppleLoginButton";
-export { default as GuestLoginButton } from "./GuestLoginButton";
