+++ src/pages/SignInPage.jsx
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 
 import {
-    AppleLoginButton,
-    GuestLoginButton,
-    KakaoLoginButton,
-} from "../domain/member/components";
+    AppleSignInButton,
+    GuestSignInButton,
+    KakaoSignInButton,
+} from "../domain/auth/components";
 import { BookClipImage, BrandLogoIWithTextImage } from "../mds/image";
 
 const LoginPage = () => {
@@ -15,9 +15,9 @@ const LoginPage = () => {
                 <BrandLogoIWithTextImage />
             </BrandLogoContainer>
             <ButtonContainer>
-                <KakaoLoginButton />
-                <AppleLoginButton />
-                <GuestLoginButton />
+                <KakaoSignInButton />
+                <AppleSignInButton />
+                <GuestSignInButton />
             </ButtonContainer>
         </Container>
     );
