+++ src/pages/index.js
@@ -10,4 +10,4 @@ export { default as AgreementPersonalPage } from "./AgreementPersonalPage";
 export { default as SettingPersonalPage } from "./SettingPersonalPage";
 export { default as TermOfServicePage } from "./TermOfServicePage";
 export { default as AgreementPage } from "./AgreementPage";
-export { default as LoginPage } from "./LoginPage";
+export { default as SignInPage } from "./SignInPage";
